+++ functions_example/CS/functions_example.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30523.141
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "functions_example", "functions_example.csproj", "{B3CC8667-2D9B-4494-973E-C716730C65C4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B3CC8667-2D9B-4494-973E-C716730C65C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B3CC8667-2D9B-4494-973E-C716730C65C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B3CC8667-2D9B-4494-973E-C716730C65C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B3CC8667-2D9B-4494-973E-C716730C65C4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F94F9044-BA85-47E1-8EE7-702DC4EC76DD}
+	EndGlobalSection
+EndGlobal
